Move to size_t where practical

+++ include/util.h
@@ -1,8 +1,10 @@
 #ifndef LIB_H
 #define LIB_H
+#include <stddef.h>
+
 void die(const char *message);
 
-char *formatBytes(unsigned bytes, float *rounded);
+char *formatBytes(size_t bytes, float *rounded);
 
-int intlen(unsigned i);
+int intlen(size_t i);
 #endif
